added interfaces to json
