+++ src/components/events/eventBox.tsx
@@ -1,23 +1,27 @@
 import React from "react"
 import events from "./events.json"
 
-let eventArray :object[];
-eventArray = events;
+interface eventCheck {
+    Title: string;
+    Date: string;
+    Description: string;
+}
+
+let eventArray :eventCheck[] = events;
 
 
 const EventsBox = () => (
     <div>
-        <p>
-            {eventArray.map((value, index) => {
-                return (
-                <p key={index}>
-                    <h4>{value.Title}</h4>
-                    <li>Date: {value.Date}</li>
-                    <li>Description: {value.Description}</li></p>
-                )
-            })}
-        </p>
-    </div>   
+        {eventArray.map(value => {
+            return (
+            <div>
+                <h4>{value.Title}</h4>
+                <li>Date: {value.Date}</li>
+                <li>Description: {value.Description}</li>
+            </div>
+            )
+        })}
+    </div>  
 )
 
 export default EventsBox;
